// frontend/src/client/types.gen.ts
export type ContractPublic = {
    contract_number?: (string | null);
    contract_name?: (string | null);
    amount?: (number | null);
    sign_date?: (string | null);
    effective_date?: (string | null);
    expiry_date?: (string | null);
    id: string;
    project_id: string;
    file_path?: (string | null);
    parsed_data?: ({
    [key: string]: unknown;
} | null);
    created_at?: (string | null);
};